+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 1.0.0 2020/12/24
 
 - Add sizeof and display function
-- Add type array (e.g. BYTE[3], float[4x4] )
+- Add type array (e.g. BYTE[3], float[4x4])
+- Support array nesting (e.g. double[4][2])
+- No longer supports string attributes (e.g. 'char' 'DWORD')
 
 ## 0.1.0 2020/12/23